requirements.txt
Flask==3.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
markitdown[all]
celery==5.4.0
redis==5.0.7
python-dotenv==1.0.1
google-cloud-storage==2.17.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
google-cloud-documentai==2.29.0